chore(br-utils): split helper functions in their own file
